requirements.txt
torch==2.0.1
numpy==1.25.0
tqdm==4.65.0
transformers==4.45.1
lightning==2.0.9
pillow==9.5.0
tokenizers==0.14
diffusers==0.30.3
torchvision